# src/Otus.Highload.Api/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "OTUS Highload Architect",
    "version" : "1.2.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/friend/set/{user_id}" : {
      "put" : {
        "parameters" : [ {
          "name" : "user_id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/UserId"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Пользователь успешно указал своего друга"
          },
          "400" : {
            "description" : "Невалидные данные ввода"
          },
          "401" : {
            "description" : "Неавторизованный доступ"
          },
          "500" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          },
          "503" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/friend/delete/{user_id}" : {
      "put" : {
        "parameters" : [ {
          "name" : "user_id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/UserId"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Пользователь успешно удалил из друзей пользователя"
          },
          "400" : {
            "description" : "Невалидные данные ввода"
          },
          "401" : {
            "description" : "Неавторизованный доступ"
          },
          "500" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          },
          "503" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/post/create" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/post_create_body"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Успешно создан пост",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostId"
                }
              }
            }
          },
          "400" : {
            "description" : "Невалидные данные ввода"
          },
          "401" : {
            "description" : "Неавторизованный доступ"
          },
          "500" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          },
          "503" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/post/update" : {
      "put" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/post_update_body"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Успешно изменен пост"
          },
          "400" : {
            "description" : "Невалидные данные ввода"
          },
          "401" : {
            "description" : "Неавторизованный доступ"
          },
          "500" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          },
          "503" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/post/delete/{id}" : {
      "put" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/PostId"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Успешно удален пост"
          },
          "400" : {
            "description" : "Невалидные данные ввода"
          },
          "401" : {
            "description" : "Неавторизованный доступ"
          },
          "500" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          },
          "503" : {
            "description" : "Ошибка сервера",
            "headers" : {
              "Retry-After" : {
                "description" : "Время, через которое еще раз нужно сделать запрос",
                "required" : false,
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "integer"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_500"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserId" : {
        "type" : "string",
        "description" : "Идентификатор пользователя"
      },
...
      "PostId" : {
        "type" : "string",
        "description" : "Идентификатор поста",
        "example" : "1d535fd6-7521-4cb1-aa6d-031be7123c4d"
      },
      "PostText" : {
        "type" : "string",
        "description" : "Текст поста",
        "example" : "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Lectus mauris ultrices eros in cursus turpis massa."
      },
...
      "inline_response_500" : {
        "required" : [ "message" ],
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Описание ошибки"
          },
          "request_id" : {
            "type" : "string",
            "description" : "Идентификатор запроса. Предназначен для более быстрого поиска проблем."
          },
          "code" : {
            "type" : "integer",
            "description" : "Код ошибки. Предназначен для классификации проблем и более быстрого решения проблем."
          }
        }
      },
...
      "post_create_body" : {
        "required" : [ "text" ],
        "type" : "object",
        "properties" : {
          "text" : {
            "$ref" : "#/components/schemas/PostText"
          }
        }
      },
      "post_update_body" : {
        "required" : [ "id", "text" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/PostId"
          },
          "text" : {
            "$ref" : "#/components/schemas/PostText"
          }
        }
      },
...
    },
...
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Авторизация по токену, который был получен в методе /login",
        "scheme" : "bearer"
      }
    }
  }
}